Refactor Ceph capacity helpers and improve fill_pool_job flexibility
- Move `get_ceph_used_capacity` from ceph_helpers to cluster.py for better centralization.
- Introduce `get_ceph_df_stats` utility to reduce code duplication when querying 'ceph df'.
- Update `get_percent_used_capacity` to utilize the new centralized df stats helper.
- Modify `fill_pool_job.yaml` to allow user-defined input files via INPUT_FILE environment variable, defaulting to /dev/zero.
- Update libtest imports to reflect the relocation of capacity functions.
